+++ src/sct/testing/utilities/analyses/__init__.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: Aresys S.r.l. <info@aresys.it>
+# SPDX-License-Identifier: MIT
+
+"""
+SCT Testing - Supported Analyses
+--------------------------------
+"""
+
+from sct.testing.utilities.analyses.pool import interf, notch, pta, ra
+
+__all__ = [
+    "interf",
+    "notch",
+    "pta",
+    "ra",
+]