+++ src/sct/testing/utilities/analyses/pool/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: Aresys S.r.l. <info@aresys.it>
+# SPDX-License-Identifier: MIT
+
+"""
+Supported Analyses Pool
+-----------------------
+"""